Sección 5: Implementing Roles and Permissions in Inertia Based Applications

+++ app/Http/Controllers/StudentController.php
@@ -16,6 +16,9 @@ class StudentController extends Controller
 {
     public function index(Request $request)
     {
+        
+        Gate::authorize('student_access');
+
         $students = Student::search($request)->paginate(10);
         $classes = ClassResource::collection(Classes::all());
         
@@ -29,6 +32,9 @@ class StudentController extends Controller
 
     public function create()
     {
+        
+        Gate::authorize('student_create');
+        
         $classes = ClassResource::collection(Classes::all());
 
         return Inertia::render('Student/Create', [
@@ -38,6 +44,9 @@ class StudentController extends Controller
 
     public function store(StoreStudentRequest $request)
     {
+        
+        Gate::authorize('student_create');
+        
         Student::create($request->validated());
 
         return redirect()->route('students.index');
@@ -45,6 +54,9 @@ class StudentController extends Controller
 
     public function edit(Student $student)
     {
+        
+        Gate::authorize('student_edit');
+        
         $classes = ClassResource::collection(Classes::all());
 
         return Inertia::render('Student/Edit', [
@@ -55,6 +67,9 @@ class StudentController extends Controller
 
     public function update(UpdateStudentRequest $request, Student $student)
     {
+        
+        Gate::authorize('student_edit');
+        
         $student->update($request->validated());
 
         return redirect()->route('students.index');
@@ -62,6 +77,9 @@ class StudentController extends Controller
 
     public function destroy(Student $student)
     {
+        
+        Gate::authorize('student_delete');
+        
         $student->delete();
 
         return redirect()->route('students.index');
